+++ demo/src/source-info.js
@@ -1,7 +1,8 @@
 const rootTIFFUrl =
-  'https://vitessce-data.storage.googleapis.com/vanderbilt.images/';
+  'https://vitessce-demo-data.storage.googleapis.com/test-data/vanderbilt.images/vanderbilt.images/';
 
 const tiffInfo = {
+  isPublic: true,
   channelNames: [
     'Cy3 - Synaptopodin (glomerular)',
     'Cy5 - THP (thick limb)',
@@ -15,6 +16,8 @@ const rootZarrUrl =
   'https://vitessce-demo-data.storage.googleapis.com/test-data/vanderbilt-data/vanderbilt_mxif_ims.zarr/mxif_pyramid';
 
 const zarrInfo = {
+  isPublic: true,
+  isPyramid: true,
   channelNames: [
     'Cy3 - Synaptopodin (glomerular)',
     'Cy5 - THP (thick limb)',
@@ -24,7 +27,22 @@ const zarrInfo = {
   url: rootZarrUrl
 };
 
+const rootStaticZarrUrl =
+  'https://vitessce-data.storage.googleapis.com/0.0.24/master_release/spraggins/spraggins.ims.zarr/spraggins.ims.zarr';
+
+const staticInfo = {
+  isPublic: false,
+  initialViewState: {
+    zoom: -1,
+    target: [1000, 500]
+  },
+  minZoom: 0,
+  channelNames: ['703.5722'],
+  url: rootStaticZarrUrl
+};
+
 export default {
   zarr: zarrInfo,
-  tiff: tiffInfo
+  tiff: tiffInfo,
+  static: staticInfo
 };