+++ src/layers/VivViewerLayer/VivViewerLayer.js
@@ -1,100 +1,38 @@
 import { CompositeLayer } from '@deck.gl/core';
 // eslint-disable-next-line import/extensions
 import VivViewerLayerBase from './VivViewerLayerBase';
-import {
-  padWithDefault,
-  isInTileBounds,
-  DEFAULT_COLOR_OFF,
-  MAX_COLOR_INTENSITY
-} from './utils';
-
-const MAX_SLIDERS_AND_CHANNELS = 6;
+import { isInTileBounds } from './utils';
+import { padColorsAndSliders } from '../utils';
 
 export default class VivViewerLayer extends CompositeLayer {
-  initializeState() {
-    this.state = {
-      imageWidth: 0,
-      imageHeight: 0,
-      tileSize: 0,
-      minZoom: 0
-    };
-  }
-
   // see https://github.com/uber/deck.gl/blob/master/docs/api-reference/layer.md#shouldupdatestate
   // eslint-disable-next-line class-methods-use-this
   shouldUpdateState({ changeFlags }) {
     return changeFlags.somethingChanged;
   }
 
-  updateState() {
-    if (this.props.loader) {
-      const {
-        imageWidth,
-        imageHeight,
-        tileSize,
-        minZoom
-      } = this.props.loader.vivMetadata;
-
-      this.setState({
-        minZoom,
-        imageWidth,
-        imageHeight,
-        tileSize
-      });
-    }
-  }
-
-  _overrideChannelProps() {
+  renderLayers() {
     const {
+      loader,
       sliderValues,
       colorValues,
       channelIsOn,
-      maxSliderValue
+      domain
     } = this.props;
-    const lengths = [sliderValues.length, colorValues.length];
-    if (lengths.every(l => l !== lengths[0])) {
-      throw Error('Inconsistent number of slider values and colors provided');
-    }
-
-    const colors = colorValues.map((color, i) =>
-      channelIsOn[i]
-        ? color.map(c => c / MAX_COLOR_INTENSITY)
-        : DEFAULT_COLOR_OFF
-    );
-
-    const sliders = sliderValues.map((slider, i) =>
-      channelIsOn[i] ? slider : [maxSliderValue, maxSliderValue]
-    );
-
-    // Need to pad sliders and colors with default values (required by shader)
-    const padSize = MAX_SLIDERS_AND_CHANNELS - colors.length;
-    if (padSize < 0) {
-      throw Error(`${lengths} channels passed in, but only 6 are allowed.`);
-    }
-
-    const paddedColorValues = padWithDefault(
-      colors,
-      DEFAULT_COLOR_OFF,
-      padSize
-    );
-    const paddedSliderValues = padWithDefault(
-      sliders,
-      [maxSliderValue, maxSliderValue],
-      padSize
-    );
-    const overrideValuesProps = {
-      ...this.props,
-      sliderValues: paddedSliderValues.flat(), // flatten for use on shaders
-      colorValues: paddedColorValues.flat()
-    };
-    return overrideValuesProps;
-  }
-
-  renderLayers() {
-    const { imageWidth, imageHeight, tileSize, minZoom } = this.state;
-
-    const { loader } = this.props;
-    const layerProps = this._overrideChannelProps();
+    const {
+      imageWidth,
+      imageHeight,
+      tileSize,
+      minZoom,
+      dtype
+    } = loader.vivMetadata;
+    const { paddedSliderValues, paddedColorValues } = padColorsAndSliders({
+      sliderValues,
+      colorValues,
+      channelIsOn,
+      domain,
+      dtype
+    });
     const getTileData = ({ x, y, z }) => {
       if (
         isInTileBounds({
@@ -115,13 +53,16 @@ export default class VivViewerLayer extends CompositeLayer {
       }
       return null;
     };
-    const layers = new VivViewerLayerBase({
+    const layers = new VivViewerLayerBase(this.props, {
+      id: `VivViewerLayerBase--${loader.type}`,
       imageWidth,
       imageHeight,
       tileSize,
       minZoom,
       getTileData,
-      ...this.getSubLayerProps(layerProps)
+      dtype,
+      colorValues: paddedColorValues,
+      sliderValues: paddedSliderValues
     });
     return layers;
   }