+++ src/layers/VivViewerLayer/VivViewerLayerBase.js
@@ -14,7 +14,10 @@ const defaultProps = {
   minZoom: { type: 'number', value: 0, compare: true },
   maxZoom: { type: 'number', value: 0, compare: true },
   renderSubLayers: { type: 'function', value: renderSubLayers, compare: false },
-  opacity: { type: 'number', value: 1, compare: true }
+  opacity: { type: 'number', value: 1, compare: true },
+  colormap: { type: 'string', value: '', compare: true },
+  dtype: { type: 'string', value: '<u2', compare: true },
+  domain: { type: 'array', value: [], compare: true }
 };
 
 export default class VivViewerLayerBase extends TileLayer {}