+++ src/layers/VivViewerLayer/utils.js
@@ -1,18 +1,7 @@
 import XRLayer from '../XRLayer';
 
-export const MAX_COLOR_INTENSITY = 255;
-
-export const DEFAULT_COLOR_OFF = [0, 0, 0];
-
 export function range(len) {
   return [...Array(len).keys()];
-}
-
-export function padWithDefault(arr, defaultValue, padWidth) {
-  for (let i = 0; i < padWidth; i += 1) {
-    arr.push(defaultValue);
-  }
-  return arr;
 }
 
 export function isInTileBounds({
@@ -32,24 +21,55 @@ export function isInTileBounds({
 
 export function renderSubLayers(props) {
   const {
-    bbox: { left, top, right, bottom }
+    bbox: { left, top, right, bottom },
+    x,
+    y,
+    z
+    // eslint-disable-next-line react/destructuring-assignment
   } = props.tile;
-  const { colorValues, sliderValues, tileSize, visible, opacity, data } = props;
-  const xrl =
-    // If image metadata is undefined, do not render this layer.
-    props.imageWidth &&
-    props.imageHeight &&
-    data &&
-    new XRLayer({
-      id: `XRLayer-left${left}-top${top}-right${right}-bottom${bottom}`,
-      data: null,
-      channelData: data,
-      sliderValues,
-      colorValues,
+  const {
+    colorValues,
+    sliderValues,
+    imageWidth,
+    imageHeight,
+    tileSize,
+    minZoom,
+    visible,
+    opacity,
+    data,
+    colormap,
+    dtype
+  } = props;
+  if (
+    isInTileBounds({
+      x,
+      y,
+      z: -z,
+      imageWidth,
+      imageHeight,
       tileSize,
-      bounds: [left, bottom, right, top],
-      opacity,
-      visible
-    });
-  return xrl;
+      minZoom
+    })
+  ) {
+    const xrl =
+      // If image metadata is undefined, do not render this layer.
+      imageWidth &&
+      imageHeight &&
+      new XRLayer({
+        id: `XRLayer-left${left}-top${top}-right${right}-bottom${bottom}`,
+        data: null,
+        channelData: data,
+        // See StaticImageLayer for why this is here.
+        sliderValues: colormap ? sliderValues.slice(0, 2) : sliderValues,
+        colorValues: colormap ? [] : colorValues,
+        tileSize,
+        bounds: [left, bottom, right, top],
+        opacity,
+        visible,
+        dtype,
+        colormap
+      });
+    return xrl;
+  }
+  return null;
 }