+++ src/layers/XRLayer/xr-layer-fragment-colormap.glsl
@@ -1,0 +1,71 @@
+#version 300 es
+#define SHADER_NAME xr-layer-fragment-shader
+precision highp float;
+precision highp int;
+precision highp usampler2D;
+#pragma glslify: jet = require("glsl-colormap/jet")
+#pragma glslify: hsv = require("glsl-colormap/hsv")
+#pragma glslify: hot = require("glsl-colormap/hot")
+#pragma glslify: cool = require("glsl-colormap/cool")
+#pragma glslify: spring = require("glsl-colormap/spring")
+#pragma glslify: summer = require("glsl-colormap/summer")
+#pragma glslify: autumn = require("glsl-colormap/autumn")
+#pragma glslify: winter = require("glsl-colormap/winter")
+#pragma glslify: bone = require("glsl-colormap/bone")
+#pragma glslify: copper = require("glsl-colormap/copper")
+#pragma glslify: greys = require("glsl-colormap/greys")
+#pragma glslify: yignbu = require("glsl-colormap/yignbu")
+#pragma glslify: greens = require("glsl-colormap/greens")
+#pragma glslify: yiorrd = require("glsl-colormap/yiorrd")
+#pragma glslify: bluered = require("glsl-colormap/bluered")
+#pragma glslify: rdbu = require("glsl-colormap/rdbu")
+#pragma glslify: picnic = require("glsl-colormap/picnic")
+#pragma glslify: rainbow = require("glsl-colormap/rainbow")
+#pragma glslify: portland = require("glsl-colormap/portland")
+#pragma glslify: blackbody = require("glsl-colormap/blackbody")
+#pragma glslify: earth = require("glsl-colormap/earth")
+#pragma glslify: electric = require("glsl-colormap/electric")
+#pragma glslify: alpha = require("glsl-colormap/alpha")
+#pragma glslify: viridis = require("glsl-colormap/viridis")
+#pragma glslify: inferno = require("glsl-colormap/inferno")
+#pragma glslify: magma = require("glsl-colormap/magma")
+#pragma glslify: plasma = require("glsl-colormap/plasma")
+#pragma glslify: warm = require("glsl-colormap/warm")
+#pragma glslify: rainbow-soft = require("glsl-colormap/rainbow-soft")
+#pragma glslify: bathymetry = require("glsl-colormap/bathymetry")
+#pragma glslify: cdom = require("glsl-colormap/cdom")
+#pragma glslify: chlorophyll = require("glsl-colormap/chlorophyll")
+#pragma glslify: density = require("glsl-colormap/density")
+#pragma glslify: freesurface-blue = require("glsl-colormap/freesurface-blue")
+#pragma glslify: freesurface-red = require("glsl-colormap/freesurface-red")
+#pragma glslify: oxygen = require("glsl-colormap/oxygen")
+#pragma glslify: par = require("glsl-colormap/par")
+#pragma glslify: phase = require("glsl-colormap/phase")
+#pragma glslify: salinity = require("glsl-colormap/salinity")
+#pragma glslify: temperature = require("glsl-colormap/temperature")
+#pragma glslify: turbidity = require("glsl-colormap/turbidity")
+#pragma glslify: velocity-blue = require("glsl-colormap/velocity-blue")
+#pragma glslify: velocity-green = require("glsl-colormap/velocity-green")
+#pragma glslify: cubehelix = require("glsl-colormap/cubehelix")
+
+
+// our texture
+uniform usampler2D channelColormap;
+
+// range
+uniform vec2 sliderValues;
+
+// opacity
+uniform float opacity;
+
+in vec2 vTexCoord;
+
+out vec4 color;
+
+
+
+void main() {
+  float intensityValue = (float(texture(channelColormap, vTexCoord).r) - sliderValues[0]) / sliderValues[1];
+
+  color = vec4(colormap(min(1.0,intensityValue)).xyz, opacity);
+}