+++ tests/XRLayer.spec.js
@@ -43,7 +43,8 @@ test('XRLayer', t => {
         0
       ],
       tileSize: 2,
-      channelData: [[0, 2, 1, 2]]
+      channelData: [new Uint32Array([0, 2, 1, 2])],
+      dtype: '<u4'
     },
     onBeforeUpdate: ({ testCase }) => t.comment(testCase.title)
   });