+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Added
 
+- StaticImageLayer added for viewing non-tiled images
+- Single channel colormaps for tiled and static
 - Expose opacity and visibility
 - Testing for layers added
 - More automated build process